+++ core/src/main/java/com/arnyminerz/escalaralcoiaicomtat/core/ui/Fonts.kt
@@ -2,6 +2,31 @@ package com.arnyminerz.escalaralcoiaicomtat.core.ui
 
 import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontStyle
+import androidx.compose.ui.text.font.FontWeight
 import com.arnyminerz.escalaralcoiaicomtat.core.R
 
-val cabinFamily = FontFamily(Font(R.font.cabin))
+val CabinFamily = FontFamily(Font(R.font.cabin))
+
+val PoppinsFamiliy = FontFamily(
+    Font(R.font.poppins_extra_light_italic, FontWeight.ExtraLight, FontStyle.Italic),
+    Font(R.font.poppins_extra_light, FontWeight.ExtraLight, FontStyle.Normal),
+
+    Font(R.font.poppins_light_italic, FontWeight.Light, FontStyle.Italic),
+    Font(R.font.poppins_light, FontWeight.Light, FontStyle.Normal),
+
+    Font(R.font.poppins_italic, FontWeight.Normal, FontStyle.Italic),
+    Font(R.font.poppins_regular, FontWeight.Normal, FontStyle.Normal),
+
+    Font(R.font.poppins_medium_italic, FontWeight.Medium, FontStyle.Italic),
+    Font(R.font.poppins_medium, FontWeight.Medium, FontStyle.Normal),
+
+    Font(R.font.poppins_bold_italic, FontWeight.Bold, FontStyle.Italic),
+    Font(R.font.poppins_bold, FontWeight.Bold, FontStyle.Normal),
+
+    Font(R.font.poppins_black_italic, FontWeight.Black, FontStyle.Italic),
+    Font(R.font.poppins_black, FontWeight.Black, FontStyle.Normal),
+
+    Font(R.font.poppins_extra_bold_italic, FontWeight.ExtraBold, FontStyle.Italic),
+    Font(R.font.poppins_extra_bold, FontWeight.ExtraBold, FontStyle.Normal),
+)
